requirements.txt
attrs==21.4.0
binpacking==1.5.2
dataclasses==0.8
dtw-python==1.3.0
fastdtw==0.3.4
gdown==4.4.0
gym==0.21
ifcfg==0.22
jsonlines==3.0.0
lmdb==1.3.0
matplotlib==3.3.4
moviepy==1.0.3
msgpack_numpy==0.4.7.1
networkx==2.5.1
numpy==1.19.5
Opencv-python==4.5.5.64
pre-commit==2.17.0
imageio==2.15.0
imageio-ffmpeg==0.4.5
imutils==0.5.4
numpy-quaternion==2021.8.30.10.33.11
scipy==1.5.3
torch==1.10.2
torch-scatter==2.0.9
torchvision==0.2.2.post3
tqdm==4.63.1
tensorflow==1.13.1
tb-nightly
yacs==0.1.8
webdataset==0.1.40
